% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Wrapper to dplyr:filter that automatically drops unused factor levels}
\usage{
filter_rows(...)
}
\arguments{
\item{...}{additional arguments that would normally be passed to dplyr #' ::filter}
}
\description{
Wrapper to dplyr:filter that automatically drops unused factor levels
}
\examples{

}
