% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_rbind.R
\name{fuzzy_rbind}
\alias{fuzzy_rbind}
\title{Title fuzzy_rbind}
\usage{
fuzzy_rbind(df1, df2)
}
\arguments{
\item{df1}{1st dataframe}

\item{df2}{2nd dataframe}
}
\value{
a dataframe with the same columns as df1
}
\description{
Use rbind when two dataframes have different columns. Extra columns from the 2nd dataframe are deleted so it matches the 1st dataframe
}
\examples{
data(iris)
iris2 <- iris[, c("Sepal.Length", "Sepal.Width", "Species")]
fuzzy_rbind(iris2, iris)

}
