% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_generics.R
\name{Length}
\alias{Length}
\title{Wrapper around length that removes NAs}
\usage{
Length(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector of length one
}
\description{
Wrapper around length that removes NAs
}
\examples{
data(iris)
Length(iris$Sepal.Length)
}
