% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_generics.R
\name{Mean}
\alias{Mean}
\title{Wrapper around mean that removes NAs}
\usage{
Mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector of length one
}
\description{
Wrapper around mean that removes NAs
}
\examples{
data(iris)
Mean(iris$Sepal.Length)
}
