% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_generics.R
\name{Median}
\alias{Median}
\title{Wrapper around median that removes NAs}
\usage{
Median(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector of length one
}
\description{
Wrapper around median that removes NAs
}
\examples{
data(iris)
Median(iris$Sepal.Length)
}
