% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_0.R
\name{prefix_0}
\alias{prefix_0}
\title{Add a prefixing zero to scalars of length 1}
\usage{
prefix_0(x)
}
\arguments{
\item{x}{A vector of length 1 (scalar)}
}
\value{
A vector of some length as x
}
\description{
Any elements with more than one character are left blank
}
\examples{
example_vector <- c(3, 7, 24, 96, 8)
sapply(example_vector, prefix_0)
}
