% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_large.R
\name{read.large}
\alias{read.large}
\title{Function to read large data files}
\usage{
read.large(...)
}
\arguments{
\item{...}{Additional arguments to be passed to read.csv}
}
\value{
A dataframe that is a representation of the data contained
 in the file
}
\description{
Function to read large data files
}
\examples{
\dontrun{
read_large(my_data.csv)
}
}
