% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip.R
\name{read_clip}
\alias{read_clip}
\title{read_clip}
\usage{
read_clip(...)
}
\arguments{
\item{...}{Additional arguments to be passed to 'read.delim'}
}
\value{
A dataframe showing the cells on the clipboard
}
\description{
Platform agnostic method for copying data from the clipboard into a dataframe. NOTE: linux version currently not
running
}
