% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_generics.R
\name{Sum}
\alias{Sum}
\title{Wrapper to sum that removes NAs}
\usage{
Sum(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
A vector of length 1
}
\description{
Wrapper to sum that removes NAs
}
\examples{
data(iris)
Sum(iris$Sepal.Length)
}
