% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_space.R
\name{k_space}
\alias{k_space}
\title{Title}
\usage{
k_space(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
data(iris)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
geom_point() +
scale_xy_continuous(labels = k_space)
}
