% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_large.R
\name{read.large}
\alias{read.large}
\title{Function to read large data files}
\usage{
read.large(...)
}
\arguments{
\item{...}{Additional arguments to be passed to read.csv}
}
\description{
Function to read large data files
}
