% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_saf.R
\name{read_saf}
\alias{read_saf}
\title{Wrapper to read.csv that does not convert strings to factors}
\usage{
read_saf(...)
}
\arguments{
\item{...}{Additional arguments to be passed to read.csv}
}
\description{
Wrapper to read.csv that does not convert strings to factors
}
